requirements.txt
openai==1.93.0
pandas==2.3.0
pdf2image==1.17.0
pdfplumber==0.11.7
Pillow==11.2.1
PyPDF2==3.0.1
pytesseract==0.3.13
python_magic==0.4.27
Requests==2.32.4
streamlit==1.46.0
torch==2.6.0+cu124
transformers==4.52.4
